LayoutTests/platform/chromium/fast/forms/suggestion-picker/date-suggestion-picker-min-max-attribute.html

<!DOCTYPE html>
<html>
<head>
<script src="../../../../../fast/js/resources/js-test-pre.js"></script>
<script src="suggestion-picker-common.js"></script>
</head>
<body style="background-color: #bbbbbb;">
<p id="description"></p>
<div id="console"></div>
<input type=date id=date value="2012-12-24" min="2012-01-02" max="2012-01-08" list=suggestions>
<datalist id=suggestions>
    <option>2012-01-01</option>
    <option>2012-01-02</option>
    <option>2012-01-03</option>
    <option>2012-01-04</option>
    <option>2012-01-05</option>
    <option>2012-01-06</option>
    <option>2012-01-07</option>
    <option>2012-01-08</option>
    <option>2012-01-09</option>
    <option>2012-01-10</option>
</datalist>
<script>
description('Tests that min/max attributes filter out suggestions as expected.');

debug('Check that page popup doesn\'t exist at first.');
shouldBeNull('document.getElementById("mock-page-popup")');

window.onload = function() {
    openPicker(document.getElementById('date'), test);
};

function test() {
    popupWindow.focus();

    shouldBe('entryValues().length', '8');
    shouldBeEqualToString('entryValues()[0]', '2012-01-02');
    shouldBeEqualToString('entryValues()[1]', '2012-01-03');
    shouldBeEqualToString('entryValues()[2]', '2012-01-04');
    shouldBeEqualToString('entryValues()[3]', '2012-01-05');
    shouldBeEqualToString('entryValues()[4]', '2012-01-06');
    shouldBeEqualToString('entryValues()[5]', '2012-01-07');
    shouldBeEqualToString('entryValues()[6]', '2012-01-08');
    shouldBeEqualToString('entryValues()[7]', '@openCalendarPicker');

    finishJSTest();
}

</script>
<script src="../../../../../fast/js/resources/js-test-post.js"></script>
</body>
</html>
