LayoutTests/platform/chromium/fast/forms/suggestion-picker/date-suggestion-picker-min-max-attribute-expected.txt

Tests that min/max attributes filter out suggestions as expected.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Check that page popup doesn't exist at first.
PASS document.getElementById("mock-page-popup") is null
PASS entryValues().length is 8
PASS entryValues()[0] is "2012-01-02"
PASS entryValues()[1] is "2012-01-03"
PASS entryValues()[2] is "2012-01-04"
PASS entryValues()[3] is "2012-01-05"
PASS entryValues()[4] is "2012-01-06"
PASS entryValues()[5] is "2012-01-07"
PASS entryValues()[6] is "2012-01-08"
PASS entryValues()[7] is "@openCalendarPicker"
PASS successfullyParsed is true

TEST COMPLETE
 
